Move query tool components to directory.
