+++ src/components/querytool/MeasureSearchCombo.js
